# docs/swagger/swagger.yaml
basePath: /
definitions:
...
  model.ChangePasswordRequest:
    properties:
      new_password:
        example: senhaNova456
        minLength: 8
        type: string
      old_password:
        example: senhaAntiga123
        type: string
    required:
    - new_password
    - old_password
    type: object
...
  model.ErrorResponse:
    properties:
      error:
        example: mensagem de erro
        type: string
    type: object
...
  model.LoginRequest:
    properties:
      email:
        example: user@example.com
        type: string
      password:
        example: senha123
        type: string
    required:
    - email
    - password
    type: object
...
  model.MessageResponse:
    properties:
      message:
        example: operação realizada com sucesso
        type: string
    type: object
...
  model.UpdateProfileRequest:
    properties:
      cellphone:
        example: "11999998888"
        type: string
      tax:
        example: 123.456.789-00
        type: string
      user_name:
        example: João Silva
        type: string
    required:
    - user_name
    type: object
  model.UpdateUserSiteFiltersRequest:
    properties:
      target_words:
        items:
          type: string
        type: array
    type: object
...
  model.UserSiteRequest:
    properties:
      site_id:
        type: integer
      target_words:
        items:
          type: string
        type: array
    type: object
  model.ValidateCheckoutRequest:
    properties:
      email:
        example: user@example.com
        type: string
      tax:
        example: 123.456.789-00
        type: string
    required:
    - email
    - tax
    type: object
  model.ValidateCheckoutResponse:
    properties:
      email_exists:
        example: false
        type: boolean
      tax_exists:
        example: false
        type: boolean
    type: object
host: localhost:8080
info:
  contact: {}
  description: API para plataforma de scraping e matching de vagas de emprego.
  title: ScrapJobs API
  version: "1.0"
paths:
...
  /api/logout:
    post:
      description: Remove o cookie de autenticacao
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.MessageResponse'
      security:
      - CookieAuth: []
      summary: Logout do usuario
      tags:
      - Auth
...
  /api/user/change-password:
    post:
      consumes:
      - application/json
      description: Altera a senha do usuario autenticado
      parameters:
      - description: Senhas antiga e nova
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/model.ChangePasswordRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.MessageResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/model.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.ErrorResponse'
      security:
      - CookieAuth: []
      summary: Alterar senha
      tags:
      - User
  /api/user/profile:
    patch:
      consumes:
      - application/json
      description: Atualiza nome, telefone e CPF do usuario
      parameters:
      - description: Dados do perfil
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/model.UpdateProfileRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.MessageResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/model.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.ErrorResponse'
      security:
      - CookieAuth: []
      summary: Atualizar perfil
      tags:
      - User
  /api/users/validate-checkout:
    post:
      consumes:
      - application/json
      description: Verifica se email e CPF ja estao cadastrados
      parameters:
      - description: Email e CPF
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/model.ValidateCheckoutRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.ValidateCheckoutResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.ErrorResponse'
      summary: Validar dados de checkout
      tags:
      - Auth
...
  /login:
    post:
      consumes:
      - application/json
      description: Autentica o usuario e retorna JWT em cookie HTTP-only
      parameters:
      - description: Credenciais de login
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/model.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Cookie Authorization definido
          schema:
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.ErrorResponse'
      summary: Login do usuario
      tags:
      - Auth
...
  /userSite:
    post:
      consumes:
      - application/json
      description: Adiciona inscricao do usuario em um site de carreiras
      parameters:
      - description: ID do site e palavras-chave
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/model.UserSiteRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/model.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.ErrorResponse'
      security:
      - CookieAuth: []
      summary: Inscrever-se em site
      tags:
      - UserSite
  /userSite/{siteId}:
    delete:
      description: Remove inscricao do usuario em um site de carreiras
      parameters:
      - description: ID do site
        in: path
        name: siteId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.MessageResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/model.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.ErrorResponse'
      security:
      - CookieAuth: []
      summary: Cancelar inscricao em site
      tags:
      - UserSite
    patch:
      consumes:
      - application/json
      description: Atualiza as palavras-chave de filtragem para um site inscrito
      parameters:
      - description: ID do site
        in: path
        name: siteId
        required: true
        type: string
      - description: Palavras-chave
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/model.UpdateUserSiteFiltersRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.MessageResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/model.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.ErrorResponse'
      security:
      - CookieAuth: []
      summary: Atualizar filtros do site
      tags:
      - UserSite
securityDefinitions:
  CookieAuth:
    in: cookie
    name: Authorization
    type: apiKey
swagger: "2.0"